requirements.txt
streamlit
psycopg2-binary
pandas
python-dotenv
reportlab
